Feature: Dynamicly allocate memory
Dynamicly allocate memory for the logging-string we
receive. The previous version, which uses a character
array with a fixed length on the stack, was vulnerable
to stack overflows when logging big strings (8k+ chars).

Now it will just allocate enough memory for whatever string
you pass into the logging functions.